Data Count Get

+++ src/main/java/lk/easy/rental/service/CustomerService.java
@@ -20,5 +20,5 @@ public interface CustomerService {
     CustomerDTO searchCustomer(String id);
     List<CustomerDTO> getAllCustomer();
     String generateCustomerIds();
-    int countCustomer();
+    long countCustomer();
 }
